Pulled amber tests of ti estimators into existing ti_estimators tests.
