+++ src/alchemlyb/tests/test_ti_estimators.py
@@ -6,8 +6,10 @@ import pytest
 import pandas as pd
 
 from alchemlyb.parsing import gmx
+from alchemlyb.parsing import amber
 from alchemlyb.estimators import TI
 import alchemtest.gmx
+import alchemtest.amber
 
 
 def gmx_benzene_coul_dHdl():
@@ -26,11 +28,29 @@ def gmx_benzene_vdw_dHdl():
 
     return dHdl
 
+def amber_simplesolvated_charge_dHdl():
+    dataset = alchemtest.amber.load_simplesolvated()
+
+    dHdl = pd.concat([amber.extract_dHdl(filename)
+                      for filename in dataset['data']['charge']])
+
+    return dHdl
+
+def amber_simplesolvated_vdw_dHdl():
+    dataset = alchemtest.amber.load_simplesolvated()
+
+    dHdl = pd.concat([amber.extract_dHdl(filename)
+                      for filename in dataset['data']['vdw']])
+
+    return dHdl
+
 
 class TIestimatorMixin:
 
     @pytest.mark.parametrize('X_delta_f', ((gmx_benzene_coul_dHdl(), 3.089),
-                                           (gmx_benzene_vdw_dHdl(), -3.056)))
+                                           (gmx_benzene_vdw_dHdl(), -3.056),
+                                           (amber_simplesolvated_charge_dHdl(), -60.114),
+                                           (amber_simplesolvated_vdw_dHdl(), 3.824)))
     def test_get_delta_f(self, X_delta_f):
         est = self.cls().fit(X_delta_f[0])
         delta_f = est.delta_f_.iloc[0, -1]
